"auto"  replaced by "T"
Hello,
Thanks for sharing this file, i'm currently working with it in my multithread code. As my code was slow i wanted to try to cout<<val; but as it was an "auto" value declaration i was wondering why it wasn't a fixed type. That being said i saw that this "auto" value declaration can be replaced by "T" as it makes the code more intelligible!